+++ doc/source/changelog.rst
@@ -3,7 +3,9 @@ Changelog
 
 next
 ----
-* Fixed `typing.Optional[bool]` to still be treated as a flag.
+* Fixed ``typing.Optional[bool]`` to still be treated as a flag.
+* Added support for ``tuple[X, ...]`` and ``Collection[X]`` (Thanks to
+  @neelmraman).
 
 6.2.0 (2021-11-24)
 ------------------